refactor: Move watcher integration tests to separate directory
