+++ watcher/tests/integration/test_auto_upload_flow.py
