+++ watcher/tests/integration/test_error_paths.py
