+++ watcher/tests/integration/test_instrument_lifecycle.py
