+++ watcher/tests/integration/test_watcher_registration.py
